requirements.txt
matplotlib==3.5.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.11.0
torch==1.12.1
torchvision==0.13.1
